Adição de dados da API EONET em Eventos Naturais

+++ app/src/main/java/com/example/cyberspace_info/listaeventosnaturais/model/GeometryEventModel.kt
@@ -5,5 +5,5 @@ import kotlin.collections.ArrayList
 
 data class GeometryEventModel(
     val date: String,
-    val coordinates : Array<Int>
+    val coordinates : Array<Double>
 )
